+++ include/skyr/core/url_serialize.hpp
@@ -1,4 +1,4 @@
-// Copyright 2018 Glyn Matthews.
+// Copyright 2018-20 Glyn Matthews.
 // Distributed under the Boost Software License, Version 1.0.
 // (See accompanying file LICENSE_1_0.txt or copy at
 // http://www.boost.org/LICENSE_1_0.txt)
@@ -6,7 +6,7 @@
 #ifndef SKYR_URL_SERIALIZE_INC
 #define SKYR_URL_SERIALIZE_INC
 
-#include <skyr/url/url_record.hpp>
+#include <skyr/core/url_record.hpp>
 
 namespace skyr {
 inline namespace v1 {
