+++ include/skyr/url.hpp
@@ -16,11 +16,11 @@
 #include <tl/expected.hpp>
 #include <skyr/config.hpp>
 #include <skyr/version.hpp>
-#include <skyr/url/url_record.hpp>
-#include <skyr/url/url_error.hpp>
+#include <skyr/core/url_record.hpp>
+#include <skyr/core/url_error.hpp>
 #include <skyr/url/url_search_parameters.hpp>
-#include <skyr/url/ipv4_address.hpp>
-#include <skyr/url/ipv6_address.hpp>
+#include <skyr/network/ipv4_address.hpp>
+#include <skyr/network/ipv6_address.hpp>
 #include <skyr/url/details/to_bytes.hpp>
 
 #if defined(SKYR_PLATFORM_MSVC)
