+++ src/core/url_parse_impl.hpp
@@ -1,4 +1,4 @@
-// Copyright 2018-19 Glyn Matthews.
+// Copyright 2018-20 Glyn Matthews.
 // Distributed under the Boost Software License, Version 1.0.
 // (See accompanying file LICENSE_1_0.txt or copy at
 // http://www.boost.org/LICENSE_1_0.txt)
@@ -10,7 +10,7 @@
 #include <system_error>
 #include <optional>
 #include <tl/expected.hpp>
-#include <skyr/url/url_record.hpp>
+#include <skyr/core/url_record.hpp>
 
 namespace skyr {
 inline namespace v1 {
