+++ src/core/url_parser_context.hpp
@@ -10,8 +10,8 @@
 #include <string_view>
 #include <optional>
 #include <tl/expected.hpp>
-#include <skyr/url/url_error.hpp>
-#include <skyr/url/url_record.hpp>
+#include <skyr/core/url_error.hpp>
+#include <skyr/core/url_record.hpp>
 #include "url_parse_impl.hpp"
 
 namespace skyr {
