+++ src/core/url_record.cpp
@@ -3,7 +3,7 @@
 // (See accompanying file LICENSE_1_0.txt or copy at
 // http://www.boost.org/LICENSE_1_0.txt)
 
-#include <skyr/url/url_record.hpp>
+#include <skyr/core/url_record.hpp>
 #include "url_schemes.hpp"
 
 namespace skyr {
