+++ src/core/url_schemes.cpp
