+++ src/filesystem/path.cpp
@@ -3,7 +3,7 @@
 // (See accompanying file LICENSE_1_0.txt or copy at
 // http://www.boost.org/LICENSE_1_0.txt)
 
-#include "skyr/url/filesystem.hpp"
+#include "skyr/filesystem/path.hpp"
 #include "skyr/url/percent_encoding/percent_decode_range.hpp"
 
 namespace skyr {
