+++ src/network/ipv4_address.cpp
@@ -11,7 +11,7 @@
 #include <optional>
 #include <cstdlib>
 #include <cerrno>
-#include <skyr/url/ipv4_address.hpp>
+#include <skyr/network/ipv4_address.hpp>
 
 namespace skyr {
 inline namespace v1 {
