+++ src/string/algorithms.hpp
