+++ src/unicode/domain.cpp
@@ -9,7 +9,7 @@
 #include <skyr/unicode/ranges/transforms/u32_transform.hpp>
 #include <skyr/unicode/domain.hpp>
 #include <skyr/unicode/idna.hpp>
-#include "url/algorithms.hpp"
+#include "string/algorithms.hpp"
 
 namespace skyr {
 inline namespace v1 {
